refactor state management
