+++ packages/slice-machine/package.json
@@ -8,7 +8,7 @@
     "postinstall": "node postinstall",
     "prepublishOnly": "next build && next export",
     "serve": "node -r esm ./server/index.js",
-    "dev-server": "PORT=9999 CWD=./tests/project nodemon -r esm ./server/index.js --watch ./server",
+    "dev-server": "cross-env PORT=9999 CWD=./tests/project nodemon -r esm ./server/index.js --watch ./server",
     "build": "next build",
     "export": "next export",
     "start": "next start",
@@ -59,6 +59,7 @@
     "@reach/menu-button": "^0.10.5",
     "@svgr/webpack": "^5.4.0",
     "@theme-ui/color": "^0.3.1",
+    "cross-env": "^7.0.3",
     "formik": "^2.1.4",
     "jest": "^26.1.0",
     "jest-config": "^26.6.3",
