+++ packages/slice-machine/src/models/slice/actions.js
@@ -1,0 +1,3 @@
+const Actions = {
+  Rename: 'rename'
+}
