+++ packages/slice-machine/src/models/slice/variation/actions.js
@@ -1,0 +1,18 @@
+export const Actions = {
+  RemoveWidget: 'variation-remove-widget'
+}
+
+export function removeWidget(state, { id, modelFieldName, key }) {
+  return {
+    ...state,
+    variations: state.variations.map(variation => {
+      if (variation.id === id) {
+        return {
+          ...variation,
+          [modelFieldName]: variation[modelFieldName].filter(e => e.key !== key)
+        }
+      }
+      return variation
+    })
+  }
+}
